move tlc.config to tla-io, prepare publishing
