+++ .unreleased/features/maven-libraries.md
@@ -1,0 +1,1 @@
+Publish the `tla-ir` and `tla-io` Scala libraries under the `org.apalache-mc` Maven Central namespace.
