+++ passes/src/test/scala/at/forsyte/apalache/tla/pp/TestTlcConfigImporter.scala
@@ -1,11 +1,11 @@
 package at.forsyte.apalache.tla.pp
 
-import at.forsyte.apalache.infra.tlc.TlcConfigParserApalache
 import at.forsyte.apalache.io.annotations.store._
 import at.forsyte.apalache.io.lir.{PrettyWriter, TextLayout}
+import at.forsyte.apalache.io.tlc.TlcConfigParserApalache
 import at.forsyte.apalache.tla.imp.SanyImporter
-import at.forsyte.apalache.tla.lir.src.SourceStore
 import at.forsyte.apalache.tla.lir.Untyped
+import at.forsyte.apalache.tla.lir.src.SourceStore
 import at.forsyte.apalache.tla.lir.transformations.impl.IdleTracker
 import at.forsyte.apalache.tla.typecheck.{MultiTypeCheckerListener, TypeCheckerTool}
 import org.junit.runner.RunWith
