+++ tla-io/src/main/scala/at/forsyte/apalache/io/tlc/config/TlcConfig.scala
@@ -1,9 +1,9 @@
-package at.forsyte.apalache.infra.tlc.config
+package at.forsyte.apalache.io.tlc.config
 
-import at.forsyte.apalache.infra.tlc.config.ConfigModelValue.STR_PREFIX
+import at.forsyte.apalache.io.tlc.config.ConfigModelValue.STR_PREFIX
 import at.forsyte.apalache.tla.lir.{TlaType1, Typed, VarT1}
 import at.forsyte.apalache.tla.typecomp.TBuilderInstruction
-import at.forsyte.apalache.tla.types.{tla, ModelValueHandler}
+import at.forsyte.apalache.tla.types.{ModelValueHandler, tla}
 
 import scala.util.parsing.input.NoPosition
